Change `build` to `host`
As part of the migration from `conda-build` 2 to `conda-build` 3, we
have switched from using `build` to `host` when we want to specify
requirements that live in the installation prefix. For build tools that
we don't want in the installation prefix, `build` is now used. While
there is no issue using `build` here as it is treated the same as in
older `conda-build` releases to ease migration pain, we shouldn't rely
on it in the future as its behavior may change.