Require `matplotlib-base` instead of `matplotlib`
As there is no need for us to require the heavy GUI dependencies that
`matplotlib` pulls in for `mplview`, switch to using `matplotlib-base`
instead. Users are certainly free to pull in the GUI backends and use
them with `mplview`. However this should make the package lighter weight
when using web backends in environments sensitive to package size (e.g.
containers).